Mission1 Step2 1. Change package - Calculator Class - 2. Test CalculatorTest.tests 3. make delimeter method and get Text method in Calculator.class 4. summary todo list and done list in README.md
