+++ README.md
@@ -7,3 +7,17 @@
 
 ## 온라인 코드 리뷰 과정
 * [텍스트와 이미지로 살펴보는 온라인 코드 리뷰 과정](https://github.com/next-step/nextstep-docs/tree/master/codereview)
+
+## ToDo
+1. README.md 정리해보기
+2. 프로덕션 코드 패키지 위치 변경
+3. void 콜론_콤마_구분자() 테스트 재확인
+4. isMinus Method 예외 확인 ( parameter : "1,2" )
+5. 구분자 문자를 확인하는 로직 추가
+
+## Done
+1. 정리 완료
+2. 위치 변경 완료
+3. void 콜론_콤마_구분자() 테스트 재확인
+4. isMinus Method 호출 전 parameter 입력 로직 변경
+5. 구분자 문자 및 문자열 계산을 위한 Text를 확인하는 메소드 추가 및 테스트 확인
